+++ src/Symfony/Component/Form/AbstractTypeExtension.php
@@ -44,10 +44,20 @@ abstract class AbstractTypeExtension implements FormTypeExtensionInterface
      */
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
-        set_error_handler(array('Symfony\Component\Form\Test\DeprecationErrorHandler', 'handleBC'));
-        $resolver->setDefaults($this->getDefaultOptions());
-        $resolver->addAllowedValues($this->getAllowedOptionValues());
-        restore_error_handler();
+        $defaults = $this->getDefaultOptions(array());
+        $allowedTypes = $this->getAllowedOptionValues(array());
+
+        if (!empty($defaults)) {
+            trigger_error('getDefaultOptions() is deprecated since version 2.1 and will be removed in 2.3. Use setDefaultOptions() instead.', E_USER_DEPRECATED);
+
+            $resolver->setDefaults($defaults);
+        }
+
+        if (!empty($allowedTypes)) {
+            trigger_error('getAllowedOptionValues() is deprecated since version 2.1 and will be removed in 2.3. Use setDefaultOptions() instead.', E_USER_DEPRECATED);
+
+            $resolver->addAllowedValues($allowedTypes);
+        }
     }
 
     /**
@@ -60,8 +70,6 @@ abstract class AbstractTypeExtension implements FormTypeExtensionInterface
      */
     public function getDefaultOptions()
     {
-        trigger_error('getDefaultOptions() is deprecated since version 2.1 and will be removed in 2.3. Use setDefaultOptions() instead.', E_USER_DEPRECATED);
-
         return array();
     }
 
@@ -75,8 +83,6 @@ abstract class AbstractTypeExtension implements FormTypeExtensionInterface
      */
     public function getAllowedOptionValues()
     {
-        trigger_error('getAllowedOptionValues() is deprecated since version 2.1 and will be removed in 2.3. Use setDefaultOptions() instead.', E_USER_DEPRECATED);
-
         return array();
     }
 }