+++ src/Symfony/Component/Form/Test/DeprecationErrorHandler.php
@@ -2,9 +2,6 @@
 
 namespace Symfony\Component\Form\Test;
 
-use Symfony\Component\Form\FormInterface as NonTestFormInterface;
-use Symfony\Component\Form\FormEvent;
-
 class DeprecationErrorHandler
 {
     public static function handle($errorNumber, $message, $file, $line, $context)
@@ -23,14 +20,5 @@ class DeprecationErrorHandler
         }
 
         return false;
-    }
-
-    public static function getFormEvent(NonTestFormInterface $form, $data)
-    {
-        set_error_handler(array('Symfony\Component\Form\Test\DeprecationErrorHandler', 'handle'));
-        $event = new FormEvent($form, $data);
-        restore_error_handler();
-
-        return $event;
     }
 }