+++ src/Symfony/Component/Form/Tests/SimpleFormTest.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Form\Tests;
 use Symfony\Component\Form\Form;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
-use Symfony\Component\Form\Util\PropertyPath;
+use Symfony\Component\PropertyAccess\PropertyPath;
 use Symfony\Component\Form\FormConfigBuilder;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\Exception\TransformationFailedException;
@@ -648,7 +648,6 @@ class SimpleFormTest extends AbstractFormTest
         $form = $this->getBuilder()
             ->addValidator($validator)
             ->getForm();
-        restore_error_handler();
 
         $validator->expects($this->once())
             ->method('validate')
@@ -658,6 +657,8 @@ class SimpleFormTest extends AbstractFormTest
         }));
 
         $form->bind('foobar');
+
+        restore_error_handler();
     }
 
     public function testBindResetsErrors()